requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
bcrypt==4.3.0
certifi==2025.4.26
cffi==1.17.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cryptography==44.0.3
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
et_xmlfile==2.0.0
executing==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
fonttools==4.57.0
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
nest-asyncio==1.6.0
numpy==2.2.5
openpyxl==3.1.5
orjson==3.10.18
packaging==25.0
pandas==2.2.3
parso==0.8.4
passlib==1.7.4
pillow==11.2.1
platformdirs==4.3.7
prompt_toolkit==3.0.51
psutil==7.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
psycopg2-pool==1.2
pure_eval==0.2.3
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.4
pydantic-extra-types==2.10.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==26.4.0
rich==14.0.0
rich-toolkit==0.14.5
rsa==4.9.1
seaborn==0.13.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==1.4.0
stack-data==0.6.3
starlette==0.46.2
tornado==6.4.2
traitlets==5.14.3
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
ujson==5.10.0
uvicorn==0.34.2
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
